+++ src/pipecatcloud/cli/commands/agent.py
@@ -1,4 +1,3 @@
-
 import aiohttp
 import questionary
 import typer
@@ -13,10 +12,12 @@ from pipecatcloud._utils.agent_utils import handle_agent_start_error
 from pipecatcloud._utils.async_utils import synchronizer
 from pipecatcloud._utils.auth_utils import requires_login
 from pipecatcloud._utils.console_utils import console
-from pipecatcloud._utils.deploy_utils import DEPLOY_STATUS_MAP
 from pipecatcloud.cli import PIPECAT_CLI_NAME
 from pipecatcloud.cli.api import API
 from pipecatcloud.cli.config import config
+from pipecatcloud._utils.deploy_utils import (
+    DEPLOY_STATUS_MAP
+)
 
 agent_cli = typer.Typer(
     name="agent", help="Agent management", no_args_is_help=True
@@ -24,6 +25,55 @@ agent_cli = typer.Typer(
 
 
 # ----- Agent Commands -----
+
+
+@agent_cli.command(name="list", help="List agents in an organization.")
+@synchronizer.create_blocking
+@requires_login
+async def list(
+    organization: str = typer.Option(
+        None,
+        "--organization",
+        "-o",
+        help="Organization to list agents for"
+    ),
+):
+    org = organization or config.get("org")
+
+    with console.status(f"Fetching agents for organization: [bold]'{org}'[/bold]", spinner="dots"):
+        data, error = await API.agents(org=org)
+
+        print(data)
+        if error:
+            typer.Exit()
+
+        if not data or len(data) == 0:
+            console.error(
+                f"[red]No agents found for namespace / organization '{org}'[/red]\n\n"
+                f"[dim]Please deploy an agent first using[/dim] [bold cyan]{PIPECAT_CLI_NAME} deploy[/bold cyan]")
+            return typer.Exit(1)
+
+        else:
+            table = Table(show_header=True, show_lines=True, border_style="dim", box=box.SIMPLE)
+            table.add_column("Name")
+            table.add_column("Agent ID")
+            table.add_column("Active Deployment ID")
+            table.add_column("Created At")
+            table.add_column("Updated At")
+
+            for service in data:
+                table.add_row(
+                    f"[bold]{service['name']}[/bold]",
+                    service['id'],
+                    service['activeDeploymentId'],
+                    service['createdAt'],
+                    service['updatedAt']
+                )
+
+            console.success(
+                table,
+                title=f"Agents for organization: {org}",
+                title_extra=f"{len(data)} results")
 
 
 @agent_cli.command(name="status", help="Get status of agent deployment")
@@ -54,6 +104,7 @@ async def status(
             console.error(f"No deployment data found for agent with name '{agent_name}'")
             return typer.Exit()
 
+        # Conditions
         conditions = data["conditions"]
         table = Table(
             show_header=True,
@@ -76,18 +127,77 @@ async def status(
                 condition['lastTransitionTime']
             )
 
+        # Deployment info
+
+        deployment_table = Table(
+            show_header=False,
+            show_lines=False,
+            box=box.SIMPLE
+        )
+        deployment_table.add_column("Key")
+        deployment_table.add_column("Value")
+        deployment_table.add_row(
+            "[bold]Image:[/bold]",
+            str(data.get("deployment", {}).get("manifest", {}).get("spec", {}).get("image", "N/A")),
+        )
+        deployment_table.add_row(
+            "[bold]Active Session Count:[/bold]",
+            str(data.get("activeSessionCount", "N/A")),
+        )
+        deployment_table.add_row(
+            "[bold]Active Deployment ID:[/bold]",
+            str(data.get("activeDeploymentId", "N/A")),
+        )
+        deployment_table.add_row(
+            "[bold]Created At:[/bold]",
+            str(data.get("createdAt", "N/A")),
+        )
+        deployment_table.add_row(
+            "[bold]Updated At:[/bold]",
+            str(data.get("updatedAt", "N/A")),
+        )
+
+        # Autoscaling info
+        autoscaling_data = data.get("autoScaling", None)
+        if autoscaling_data:
+            autoscaling_table = Table(
+                show_header=False,
+                show_lines=False,
+                box=box.SIMPLE
+            )
+            autoscaling_table.add_column("Key")
+            autoscaling_table.add_column("Value")
+
+            autoscaling_table.add_row(
+                "[bold]Max instances:[/bold]",
+                str(autoscaling_data.get("maxReplicas", 0)),
+            )
+            autoscaling_table.add_row(
+                "[bold]Min instances:[/bold]",
+                str(autoscaling_data.get("minReplicas", 0)),
+            )
+            autoscaling_panel = Panel(
+                autoscaling_table,
+                title="[bold]Scaling configuration:[/bold]",
+                title_align="left",
+                border_style="dim",
+            )
+
         color = "bold green" if data['ready'] else "bold yellow"
         subtitle = f"[dim]Start a new active session with[/dim] [bold cyan]{PIPECAT_CLI_NAME} agent start {agent_name}[/bold cyan]" if data[
             'ready'] else f"[dim]For more information check logs with[/dim] [bold cyan]{PIPECAT_CLI_NAME} agent logs {agent_name}[/bold cyan]"
         console.print(
             Panel(
                 Group(
+                    deployment_table,
+                    autoscaling_panel if autoscaling_panel else "",
+                    table,
                     Panel(
                         f"[{color}]Health: {'Ready' if data['ready'] else 'Stopped'}[/]",
                         border_style="green" if data['ready'] else "yellow",
                         expand=False,
-                    ),
-                    table),
+                    )
+                ),
                 title=f"Status for agent [bold]{agent_name}[/bold]",
                 title_align="left",
                 subtitle_align="left",
@@ -95,65 +205,51 @@ async def status(
             ))
 
 
+@agent_cli.command(name="sessions", help="List active sessions for an agent")
+@synchronizer.create_blocking
+@requires_login
+async def sessions(
+    agent_name: str,
+    organization: str = typer.Option(
+        None,
+        "--organization",
+        "-o",
+        help="Organization to list sessions for"
+    ),
+):
+    org = organization or config.get("org")
+
+    with Live(console.status(f"[dim]Looking up agent with name '{agent_name}'[/dim]", spinner="dots")) as live:
+        data, error = await API.agent(agent_name=agent_name, org=org, live=live)
+
+        live.stop()
+
+        if error:
+            return typer.Exit()
+
+        if not data:
+            console.error(f"No deployment data found for agent with name '{agent_name}'")
+            return typer.Exit()
+
+        console.print(
+            "[yellow]Please note: this method is currently work in progress and will be updated in the future with more information[/yellow]")
+
+        if data.get('activeSessionCount', 0) > 0:
+            console.success(
+                f"{data.get('activeSessionCount', 0)}",
+                title=f"Active session count for agent {agent_name} [dim]({org})[/dim]")
+        else:
+            console.error(
+                f"No active sessions found for agent {agent_name}",
+                title=f"Active session count for agent {agent_name} [dim]({org})[/dim]",
+                subtitle=f"[white dim]Start a new session with[/white dim] [bold cyan]{PIPECAT_CLI_NAME} agent start {agent_name}[/bold cyan]")
+
+
 @agent_cli.command(name="scale", help="Modify agent runtime configuration")
 @synchronizer.create_blocking
 @requires_login
 async def scale():
     console.error("Not implemented")
-
-
-@agent_cli.command(name="list", help="List agents in an organization.")
-@synchronizer.create_blocking
-@requires_login
-async def list(
-    organization: str = typer.Option(
-        None,
-        "--organization",
-        "-o",
-        help="Organization to list agents for"
-    ),
-):
-    token = config.get("token")
-    org = organization or config.get("org")
-
-    with console.status(f"Fetching agents for organization: [bold]'{org}'[/bold]", spinner="dots"):
-        async with aiohttp.ClientSession() as session:
-            response = await session.get(
-                f"{API.construct_api_url('services_path').format(org=org)}",
-                headers={"Authorization": f"Bearer {token}"},
-            )
-
-        data = await response.json()
-
-        if "error" in data:
-            console.error(
-                f"[red]Unable to get agents for '{org}'[/red]\n\n"
-                f"[dim]API response:[/dim] {data['error']}",
-            )
-        else:
-            table = Table(show_header=True, show_lines=True, border_style="dim", box=box.SIMPLE)
-            table.add_column("Name")
-            table.add_column("ID")
-            table.add_column("Active Deployment")
-            table.add_column("Created At")
-            table.add_column("Updated At")
-
-            for service in data['services']:
-                table.add_row(
-                    f"[bold]{service['name']}[/bold]",
-                    service['id'],
-                    service['activeDeploymentId'],
-                    service['createdAt'],
-                    service['updatedAt']
-                )
-
-            console.print(Panel(
-                table,
-                padding=1,
-                title=f"[bold]Agents for organization: {org}[/bold]",
-                title_align="left",
-                border_style="green",
-            ))
 
 
 @agent_cli.command(name="logs", help="Get logs for the given agent.")
@@ -194,18 +290,31 @@ async def delete(
         "-o",
         help="Organization to delete agent from",
     ),
+    force: bool = typer.Option(
+        False,
+        "--force",
+        "-f",
+        help="Bypass prompt for confirmation",
+    ),
 ):
-    token = config.get("token")
     org = organization or config.get("org")
 
-    with console.status(f"Deleting agent: [bold]'{agent_name}'[/bold]", spinner="dots"):
-        async with aiohttp.ClientSession() as session:
-            response = await session.delete(
-                f"{API.construct_api_url('services_path').format(org=org)}/{agent_name}",
-                headers={"Authorization": f"Bearer {token}"},
-            )
-            data = await response.json()
-            console.print(data)
+    if not force:
+        if not await questionary.confirm("Are you sure you want to delete this agent? Note: active sessions will not be interrupted and will continue to run until completion.").ask_async():
+            console.print("[bold]Aborting delete request[/bold]")
+            return typer.Exit(1)
+
+    with console.status(f"[dim]Deleting agent: [bold]'{agent_name}'[/bold][/dim]", spinner="dots"):
+        data, error = await API.agent_delete(agent_name=agent_name, org=org)
+
+        if error:
+            return typer.Exit(1)
+
+        if not data:
+            console.error(f"Agent '{agent_name}' not found in namespace / organization '{org}'")
+            return typer.Exit(1)
+
+        console.success(f"Agent '{agent_name}' deleted successfully")
 
 
 @agent_cli.command(name="deployments", help="Get deployments for an agent.")
@@ -226,7 +335,7 @@ async def deployments(
     error_code = None
 
     try:
-        with console.status(f"Fetching deployments for agent: [bold]'{agent_name}'[/bold]", spinner="dots"):
+        with console.status(f"[dim]Fetching deployments for agent: [bold]'{agent_name}'[/bold][/dim]", spinner="dots"):
             async with aiohttp.ClientSession() as session:
                 response = await session.get(
                     f"{API.construct_api_url('services_deployments_path').format(org=org, service=agent_name)}",
@@ -235,7 +344,9 @@ async def deployments(
             if response.status != 200:
                 error_code = str(response.status)
                 response.raise_for_status()
+
             data = await response.json()
+            print(data)
 
             table = Table(
                 show_header=True,
@@ -272,102 +383,112 @@ async def deployments(
 @synchronizer.create_blocking
 @requires_login
 async def start(
-    ctx: typer.Context,
-    agent_name: str,
+    agent_name: str = typer.Argument(
+        help="Name of the agent to start e.g. 'my-agent'"),
     force: bool = typer.Option(
         False,
         "--force",
-        "--f",
+        "-f",
         help="Bypass prompt for confirmation",
         rich_help_panel="Start Configuration",
-    ),
-    organization: str = typer.Option(
-        None,
-        "--organization",
-        "--org",
-        help="Organization to start agent for",
     ),
     api_key: str = typer.Option(
         None,
         "--api-key",
-        "--key",
+        "-k",
         help="Public API key to use for starting agent",
         rich_help_panel="Start Configuration",
     ),
     data: str = typer.Option(
         None,
         "--data",
-        "--d",
+        "-d",
         help="Data to pass to the agent (stringified JSON)",
         rich_help_panel="Start Configuration",
     ),
     use_daily: bool = typer.Option(
         False,
         "--use-daily",
-        "--daily",
+        "-D",
         help="Create a Daily WebRTC session for the agent",
         rich_help_panel="Start Configuration",
     ),
+    organization: str = typer.Option(
+        None,
+        "--organization",
+        "-o",
+        help="Organization which the agent belongs to",
+    ),
 ):
-    console = Console()
-
-    token = config.get("token")
     org = organization or config.get("org")
-    default_public_key = api_key or config.get("default_public_key")
-    default_public_key_name = "CLI provided" if api_key else config.get("default_public_key_name")
 
-    if not default_public_key:
-        print_api_error("PCC-1002", f"Unable to start agent '{agent_name}' without public api key")
+    default_public_api_key = api_key or config.get("default_public_key")
+    default_public_api_key_name = "CLI provided" if api_key else config.get(
+        "default_public_key_name")
+
+    if not default_public_api_key:
+        console.print(
+            Panel(
+                f"No public API key provided. Please provide a public API key using the --api-key flag or set a default using [bold cyan]{PIPECAT_CLI_NAME} organizations keys use[/bold cyan].\n\n"
+                f"If you have not yet created a public API key, you can do so by running [bold cyan]{PIPECAT_CLI_NAME} organizations keys create[/bold cyan].",
+                title="Public API Key Required",
+                title_align="left",
+                border_style="yellow",
+            ))
+
         return typer.Exit(1)
 
     # Confirm start request
     if not force:
         console.print(Panel(
             f"Agent Name: {agent_name}\n"
-            f"Organization: {org}\n"
-            f"Public API Key: {default_public_key_name} [dim]{default_public_key}[/dim]\n"
+            f"Public API Key: {default_public_api_key_name} [dim]{default_public_api_key}[/dim]\n"
             f"Use Daily: {use_daily}\n"
             f"Data: {data}",
             title=f"[bold]Start Request for agent: {agent_name}[/bold]",
             title_align="left",
             border_style="yellow",
         ))
-        if not await questionary.confirm("Are you sure you want to start this agent?").ask_async():
+        if not await questionary.confirm("Are you sure you want to start an active session for this agent?").ask_async():
             console.print("[bold]Aborting start request[/bold]")
             return typer.Exit(1)
 
-    # Check if agent exists and is healthy
-    with Live(console.status(f"Sending start request with key: {default_public_key_name}", spinner="dots"), refresh_per_second=4) as live:
-        error_code = None
-        try:
-            async with aiohttp.ClientSession() as session:
-                response = await session.post(
-                    f"{API.construct_api_url('start_path').format(service=agent_name)}",
-                    headers={"Authorization": f"Bearer {default_public_key}"},
-                    json={
-                        "createDailyRoom": bool(use_daily),
-                        "body": {}
-                    }
-                )
-                if response.status != 200:
-                    error_code = handle_agent_start_error(response.status)
-                    response.raise_for_status()
+    with Live(console.status(f"[dim]Checking agent health...[/dim]", spinner="dots"), refresh_per_second=4) as live:
 
-        except Exception as e:
-            live.update(console.status(f"Agent '{agent_name}' failed to start", spinner="dots"))
+        health_data, error = await API.agent(agent_name=agent_name, org=org, live=live)
+        if not health_data or not health_data['ready']:
             live.stop()
-            logger.debug(e)
-            print_api_error(
-                error_code,
-                f"Unable to start agent '{agent_name}'. Please check logs for more information.")
+            console.error(
+                f"Agent '{agent_name}' does not exist or is not in a health state. Please check the agent status with [bold cyan]{PIPECAT_CLI_NAME} agent status {agent_name}[/bold cyan]")
             return typer.Exit(1)
 
-        live.update(console.status(f"Agent '{agent_name}' started successfully", spinner="dots"))
+        live.update(
+            console.status(
+                f"[dim]Agent '{agent_name}' is healthy, sending start request...[/dim]",
+                spinner="dots"))
+
+        data, error = await API.start_agent(agent_name=agent_name, api_key=default_public_api_key, use_daily=use_daily, data=data, live=live)
+
+        if error:
+            return typer.Exit(1)
+
+        if not data:
+            live.stop()
+            console.error(
+                f"Agent '{agent_name}' not found. Have you deployed the agent?",
+                subtitle=f"[white dim]Deploy an agent with[/white dim] [bold cyan]{PIPECAT_CLI_NAME} deploy[/bold cyan]")
+            return typer.Exit(1)
+
         live.stop()
 
-        console.print(Panel(
-            f"Agent '{agent_name}' started successfully",
-            title=f"{PANEL_TITLE_SUCCESS}",
-            title_align="left",
-            border_style="green",
-        ))
+        message = f"Agent '{agent_name}' started successfully"
+        if use_daily and isinstance(data, dict):
+            daily_room = data.get("dailyRoom")
+            daily_token = data.get("dailyToken")
+            if daily_room:
+                message += f"\n\nDaily room: [link={daily_room}?t={daily_token}]{daily_room}?t={daily_token}[/link]"
+            return console.success(
+                message,
+                subtitle=f"[white dim]Join the session:[/white dim] [link={daily_room}?t={daily_token}]click here[/link]")
+
+        console.success(message)